chore: game 관련 타입을 match로 변경
