--- src/queries/useMatchById/api.ts
@@ -1,8 +1,0 @@
-import instance from '@/api/instance';
-import { GameDetailType } from '@/types/game';
-
-export const getMatchById = async (gameId: string) => {
-  const { data } = await instance.get<GameDetailType>(`/games/${gameId}`);
-
-  return data;
-};
