+++ src/queries/useMatchById/query.ts
@@ -1,20 +1,19 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import { getMatchById } from './api';
+import { getMatchById } from '@/api/match';
 
 export const QUERY_KEY = {
   MATCH_DETAIL: 'MATCH_DETAL',
 };
 
 export default function useGameDetail(matchId: string) {
-  const { data, isLoading, error } = useSuspenseQuery({
+  const { data, error } = useSuspenseQuery({
     queryKey: [QUERY_KEY.MATCH_DETAIL, matchId],
     queryFn: () => getMatchById(matchId),
   });
 
   return {
     matchDetail: data,
-    isLoading,
     error,
   };
 }
