+++ src/queries/useMatchCheerById/query.ts
@@ -1,16 +1,15 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 
-import { getMatchCheerById } from './api';
+import { getMatchCheerById } from '@/api/match';
 
 export const useMatchCheerById = (matchId: string) => {
-  const { data, isLoading, error } = useSuspenseQuery({
+  const { data, error } = useSuspenseQuery({
     queryKey: ['match-cheer', matchId],
     queryFn: () => getMatchCheerById(matchId),
   });
 
   return {
     cheers: data,
-    isLoading,
     error,
   };
 };
