--- src/queries/useMatchCheerById/type.ts
@@ -1,4 +1,0 @@
-export type MatchCheerType = {
-  gameTeamId: number;
-  cheerCount: number;
-};
